requirements.txt
Pillow
requests
thefuzz
python-Levenshtein
pymupdf
